Rename folder free_cash_flow_statement to cash_flow_statement
