+++ src/domain/financial_statements/cash_flow_statement/financing_activities.py
