+++ src/domain/financial_statements/cash_flow_statement/operating_activities.py
